domain layer 수정
