--- MemoWithTags/MemoWithTags/Domain/UseCase/Auth/LoginUseCase.swift
@@ -1,37 +1,0 @@
-//
-//  LoginUseCase.swift
-//  MemoWithTags
-//
-//  Created by 최진모 on 12/29/24.
-//
-
-protocol LoginUseCase {
-    func execute(email: String, password: String) async -> Result<Void, LoginError>
-}
-
-final class DefaultLoginUseCase: LoginUseCase {
-    private let authRepository: AuthRepository
-
-    init(authRepository: AuthRepository) {
-        self.authRepository = authRepository
-    }
-
-    func execute(email: String, password: String) async -> Result<Void, LoginError> {
-        do {
-            let dto = try await authRepository.login(email: email, password: password)
-            let auth = dto.toAuth()
-            
-            let isAccessSaved = KeyChainManager.shared.saveAccessToken(token: auth.accessToken)
-            let isRefreshSaved = KeyChainManager.shared.saveRefreshToken(token: auth.refreshToken)
-            
-            if isAccessSaved && isRefreshSaved {
-                return .success(())
-            } else {
-                return .failure(.tokenSaveError)
-            }
-        } catch let error {
-            return .failure(.from(baseError: error as! BaseError))
-        }
-    }
-}
-
